+++ engine/src/main/java/org/hibernate/search/engine/environment/classpath/spi/ClassResolver.java
@@ -6,8 +6,6 @@
  */
 package org.hibernate.search.engine.environment.classpath.spi;
 
-import org.hibernate.search.engine.environment.service.spi.Service;
-
 /**
  * A resolver of Java classes.
  *
@@ -30,8 +28,7 @@ public interface ClassResolver {
 	/**
 	 * Discovers and instantiates implementations of the named service contract.
 	 * <p>
-	 * NOTE : the term service here is used differently than {@link Service}.
-	 * Instead here we are talking about services as defined by {@link java.util.ServiceLoader}.
+	 * NOTE : We are talking about services as defined by {@link java.util.ServiceLoader}.
 	 *
 	 * @param serviceContract The java type defining the service contract
 	 * @param <T> The type of the service contract
