+++ mapper/pojo/src/main/java/org/hibernate/search/mapper/pojo/mapping/building/impl/BridgeBuildContextImpl.java
@@ -20,6 +20,6 @@ public class BridgeBuildContextImpl implements BridgeBuildContext {
 
 	@Override
 	public BeanProvider getBeanProvider() {
-		return buildContext.getServiceManager().getBeanProvider();
+		return buildContext.getBeanProvider();
 	}
 }
